requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
dicttoxml==1.7.16
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.58.1
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
xmltodict==0.14.2
